Extract loader operation to separate file
